+++ src/Column/ArrayColumn.php
@@ -5,10 +5,10 @@ declare(strict_types=1);
 namespace Yiisoft\Db\Pgsql\Column;
 
 use Yiisoft\Db\Pgsql\ArrayParser;
-use Yiisoft\Db\Schema\Column\ArrayColumnSchema as BaseArrayColumnSchema;
+use Yiisoft\Db\Schema\Column\ArrayColumn as BaseArrayColumn;
 use Yiisoft\Db\Syntax\ParserToArrayInterface;
 
-final class ArrayColumnSchema extends BaseArrayColumnSchema
+final class ArrayColumn extends BaseArrayColumn
 {
     protected function getParser(): ParserToArrayInterface
     {