+++ src/Column/BinaryColumn.php
@@ -4,14 +4,14 @@ declare(strict_types=1);
 
 namespace Yiisoft\Db\Pgsql\Column;
 
-use Yiisoft\Db\Schema\Column\BinaryColumnSchema as BaseBinaryColumnSchema;
+use Yiisoft\Db\Schema\Column\BinaryColumn as BaseBinaryColumn;
 
 use function hex2bin;
 use function is_string;
 use function str_starts_with;
 use function substr;
 
-final class BinaryColumnSchema extends BaseBinaryColumnSchema
+final class BinaryColumn extends BaseBinaryColumn
 {
     public function phpTypecast(mixed $value): mixed
     {