+++ src/Column/ColumnBuilder.php
@@ -5,52 +5,52 @@ declare(strict_types=1);
 namespace Yiisoft\Db\Pgsql\Column;
 
 use Yiisoft\Db\Constant\ColumnType;
-use Yiisoft\Db\Schema\Column\ColumnSchemaInterface;
+use Yiisoft\Db\Schema\Column\ColumnInterface;
 
 final class ColumnBuilder extends \Yiisoft\Db\Schema\Column\ColumnBuilder
 {
-    public static function boolean(): ColumnSchemaInterface
+    public static function boolean(): ColumnInterface
     {
-        return new BooleanColumnSchema(ColumnType::BOOLEAN);
+        return new BooleanColumn(ColumnType::BOOLEAN);
     }
 
-    public static function bit(int|null $size = null): ColumnSchemaInterface
+    public static function bit(int|null $size = null): ColumnInterface
     {
-        return new BitColumnSchema(ColumnType::BIT, size: $size);
+        return new BitColumn(ColumnType::BIT, size: $size);
     }
 
-    public static function tinyint(int|null $size = null): ColumnSchemaInterface
+    public static function tinyint(int|null $size = null): ColumnInterface
     {
-        return new IntegerColumnSchema(ColumnType::TINYINT, size: $size);
+        return new IntegerColumn(ColumnType::TINYINT, size: $size);
     }
 
-    public static function smallint(int|null $size = null): ColumnSchemaInterface
+    public static function smallint(int|null $size = null): ColumnInterface
     {
-        return new IntegerColumnSchema(ColumnType::SMALLINT, size: $size);
+        return new IntegerColumn(ColumnType::SMALLINT, size: $size);
     }
 
-    public static function integer(int|null $size = null): ColumnSchemaInterface
+    public static function integer(int|null $size = null): ColumnInterface
     {
-        return new IntegerColumnSchema(ColumnType::INTEGER, size: $size);
+        return new IntegerColumn(ColumnType::INTEGER, size: $size);
     }
 
-    public static function bigint(int|null $size = null): ColumnSchemaInterface
+    public static function bigint(int|null $size = null): ColumnInterface
     {
-        return new IntegerColumnSchema(ColumnType::BIGINT, size: $size);
+        return new IntegerColumn(ColumnType::BIGINT, size: $size);
     }
 
-    public static function binary(int|null $size = null): ColumnSchemaInterface
+    public static function binary(int|null $size = null): ColumnInterface
     {
-        return new BinaryColumnSchema(ColumnType::BINARY, size: $size);
+        return new BinaryColumn(ColumnType::BINARY, size: $size);
     }
 
-    public static function array(ColumnSchemaInterface|null $column = null): ColumnSchemaInterface
+    public static function array(ColumnInterface|null $column = null): ColumnInterface
     {
-        return new ArrayColumnSchema(ColumnType::ARRAY, column: $column);
+        return new ArrayColumn(ColumnType::ARRAY, column: $column);
     }
 
-    public static function structured(string|null $dbType = null, array $columns = []): ColumnSchemaInterface
+    public static function structured(string|null $dbType = null, array $columns = []): ColumnInterface
     {
-        return new StructuredColumnSchema(ColumnType::STRUCTURED, dbType: $dbType, columns: $columns);
+        return new StructuredColumn(ColumnType::STRUCTURED, dbType: $dbType, columns: $columns);
     }
 }