+++ CHANGELOG.md
@@ -22,6 +22,13 @@
 - Enh #370: Refactor `Schema::normalizeDefaultValue()` method and move it to `ColumnFactory` class (@Tigrov)
 - New #373: Override `QueryBuilder::prepareBinary()` method (@Tigrov)
 - Chg #375: Update `QueryBuilder` constructor (@Tigrov)
+- Enh #374: Use `ColumnDefinitionBuilder` to generate table column SQL representation (@Tigrov)
+- Enh #378: Improve loading schemas of views (@Tigrov)
+- Enh #379: Remove `ColumnInterface` (@Tigrov)
+- Enh #380: Rename `ColumnSchemaInterface` to `ColumnInterface` (@Tigrov)
+- Enh #381, #383: Add `ColumnDefinitionParser` class (@Tigrov)
+- Enh #382: Replace `DbArrayHelper::getColumn()` with `array_column()` (@Tigrov)
+- New #384: Add `IndexMethod` class (@Tigrov)
 
 ## 1.3.0 March 21, 2024